requirement.txt
sqlalchemy==2.0.30
pandas==2.2.2
psycopg2-binary==2.9.9
pyodbc==5.0.1
mysql-connector-python==8.3.0
alembic==1.13.1
python-dotenv==1.0.1